+++ app/components/EnhancedTable.js
@@ -29,15 +29,17 @@ const styles = theme => ({
 type Props = {
   classes: {},
   data: [],
+  orderDirection?: string,
   columnConfig: [],
   defaultOrderBy: string,
   secondarySorts: [],
   selectAll: boolean,
-  handleAddByFile: ?() => {},
-  handleAddByFolder: ?() => {},
-  getSuggestions: ?() => {},
+  multiSelections?: boolean,
   onSelectedRowIds: () => {},
-  onAutosuggestInputChanged: ?() => {}
+  handleAddByFile?: () => {},
+  handleAddByFolder?: () => {},
+  getSuggestions?: () => {},
+  onAutosuggestInputChanged?: () => {}
 };
 
 function getDataRowIds(data) {
@@ -56,14 +58,14 @@ function getAllSelectableRowIds(data) {
 class EnhancedTable extends Component<Props> {
   props: Props;
   state = {
-    order: 'asc',
+    order: this.props.orderDirection,
     orderBy: this.props.defaultOrderBy || this.props.columnConfig[0].name,
     selectedRowIds: []
   };
 
   componentWillReceiveProps(nextProps) {
     // You don't have to do this check first, but it can help prevent an unneeded render
-    if (nextProps.selectAll && nextProps.data.length &&
+    if (nextProps.multiSelections && nextProps.selectAll && nextProps.data.length &&
       (this.props.data !== nextProps.data ||
       this.state.selectedRowIds.length === 0 ||
       nextProps.selectAll !== this.props.selectAll)) {
@@ -111,13 +113,6 @@ class EnhancedTable extends Component<Props> {
     const { columnConfig } = this.props;
     const checkboxColumn = {
       name: 'checkbox',
-      header: (
-        <Checkbox
-          checked={this.areAnyChecked()}
-          onChange={this.onChangeCheckBoxHeader}
-          {...this.headerCheckBoxProps()}
-        />
-      ),
       cell: rowData => (
         !rowData.disabled ?
           <Checkbox
@@ -129,6 +124,15 @@ class EnhancedTable extends Component<Props> {
       width: 72,
       onHeaderClick: false
     };
+    if (this.props.multiSelections) {
+      const header = (
+        <Checkbox
+          checked={this.areAnyChecked()}
+          onChange={this.onChangeCheckBoxHeader}
+          {...this.headerCheckBoxProps()}
+        />);
+      checkboxColumn.header = header;
+    }
     const stringCellProps = { style: { paddingRight: 0 } };
     const numericCellProps = { numeric: true };
     const columns = columnConfig.map(c => ({
@@ -161,9 +165,15 @@ class EnhancedTable extends Component<Props> {
           selectedRowIds: prevState.selectedRowIds.filter(id => id !== rowData.id)
         };
       }
+      if (this.props.multiSelections) {
+        return {
+          selectedRowIds: [...prevState.selectedRowIds, rowData.id]
+        };
+      }
       return {
-        selectedRowIds: [...prevState.selectedRowIds, rowData.id]
+        selectedRowIds: [rowData.id]
       };
+
     }, this.reportSelectedRowIds);
   }
 
@@ -222,5 +232,14 @@ class EnhancedTable extends Component<Props> {
     );
   }
 }
+
+EnhancedTable.defaultProps = {
+  orderDirection: 'asc',
+  multiSelections: false,
+  handleAddByFile: undefined,
+  handleAddByFolder: undefined,
+  getSuggestions: undefined,
+  onAutosuggestInputChanged: undefined
+};
 
 export default withStyles(styles)(EnhancedTable);